Refactor exception handling
The previous commits worked on replacing System.err.println() with
logger.error(). However, this also showed that many methods were
printing errors directly to the log instead of throwing an exception to
the caller.

Passing an exception up the stack is a better design because it removes
all the try-catch statements and also removes the need of creating
unnecessary Loggers in small classes.