Add weight for depnn
The logp weight is no longer weight 0, but a separate weight, along
with depnn. This is to allow for future extra weights, without having
to worry about the IDs of these extra features.

The format of weights files has been changed so that logp and depnn
weights are prefixed with an identifier.